requirements.txt
alabaster==0.7.12
astroid==2.4.1
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.2
azure-core==1.7.0
azure-eventhub==5.1.0
azure-iot-device==2.1.4
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.5
bokeh==2.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
cryptography==2.9.2
dask==2.19.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distributed==2.19.0
docutils==0.16
entrypoints==0.3
flake8==3.8.1
fsspec==0.7.4
HeapDict==1.0.1
helpdev==0.7.1
idna==2.9
imagesize==1.2.0
importlib-metadata==1.7.0
intervaltree==3.0.2
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
janus==0.4.0
jedi==0.15.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
keyring==21.2.1
lazy-object-proxy==1.4.3
locket==0.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.5
numpydoc==0.9.2
packaging==20.3
paho-mqtt==1.5.0
pandas==1.0.5
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
partd==1.1.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psutil==5.7.0
ptyprocess==0.6.0
py4j==0.10.7
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.0.2
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.2
pymongo==3.10.1
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.7.2
PyQtWebEngine==5.12.1
pyrsistent==0.16.0
PySocks==1.7.1
pyspark==2.4.6
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.3
QtPy==1.9.0
requests==2.23.0
requests-unixsocket==0.2.0
rope==0.17.0
Send2Trash==1.5.0
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.3
spyder-kernels==1.9.2
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
uamqp==1.2.9
urllib3==1.25.9
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
wrapt==1.12.1
yapf==0.30.0
zict==2.0.0
zipp==3.1.0
